Moved stuff around
Former-commit-id: 472454c400e12e6dfe844c752ebe47f9332d816c [formerly 16698632af228a4327edc03f28cc042541f8ec36] [formerly 223b5337166e8d55e8ae42ce4744566e76d3501f]
Former-commit-id: 9e2802296301d2f9454893510884013375e5f992
Former-commit-id: ecd2d67e7fadbccf6359b26898951c47288f05ed

+++ misc/scripts/upgrade.sh
@@ -66,4 +66,6 @@ ${BOLD}$(cat /tmp/pacstall-up-list | tr '\n' ' ')${NORMAL}"
         exit 1
     fi
 fi
-rm -f /tmp/pacstall-up-list
+if test -f "/tmp/pacstall-up-list"; then
+    rm -f /tmp/pacstall-up-list
+fi
